python/requirements.txt
faiss-cpu
langchain-community
sentence-transformers